+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a hook (`'swimShapeChanged'`) that is called after the shape changer completed the shape changing process. It passes the new token in order to work with that further in code.
 - Dramatic Task Planner Macro. Read the Wiki on it, it has a couple features that are explained there.
 - Added `swim.revert_shape_change(token)` to revert the passed `token` back to its original form.
+- Added automatic power backlash execution for the BRSW support.
+- - Currently only automates regular backlash.
+- - Dynamic Backlash can be set in the SWIM settings but it will only roll on the table (make sure to import it into your world and set the name in the SWIM settings).
+- - This can also be turned off independently of th BRSW support.
+- - The regular backlash is fully automated and will also revert shape change forms, dismiss summoned creatures and terminate all powers if they are set with the Power Effect Builder.
 ### Changed
 - When removing maintained power, gathering effects is now in a `try` statement so that the function continues if an error occurs (i.e. when the token changed or doesn't exist anymore.)
 ### Fixed